break out data processing stages, save partitioned parquets, profile bottlenecks
