+++ daskify_rt/logs/A3_sjoin_vp_segments.log
@@ -79,3 +79,54 @@
 2023-01-07 22:47:39.441 | INFO     | __main__:compute_and_export:235 - exported 75: 0:00:21.239106
 2023-01-07 22:47:39.442 | INFO     | __main__:<module>:292 - 75: 0:00:23.633449
 2023-01-07 22:47:39.442 | INFO     | __main__:<module>:295 - execution time: 0:32:46.768594
+2023-01-10 16:30:39.756 | INFO     | __main__:<module>:229 - Analysis date: 2022-10-12
+2023-01-10 16:30:44.948 | INFO     | __main__:<module>:236 - get unique operators to loop over: 0:00:05.148442
+2023-01-10 16:33:51.406 | INFO     | __main__:<module>:229 - Analysis date: 2022-10-12
+2023-01-10 16:33:56.976 | INFO     | __main__:<module>:236 - get unique operators to loop over: 0:00:05.569562
+2023-01-10 16:37:51.805 | INFO     | __main__:<module>:229 - Analysis date: 2022-10-12
+2023-01-10 16:37:56.683 | INFO     | __main__:<module>:236 - get unique operators to loop over: 0:00:04.878255
+2023-01-10 16:38:21.771 | INFO     | __main__:<module>:276 - 110: 0:00:25.086319
+2023-01-10 16:38:47.097 | INFO     | __main__:<module>:276 - 126: 0:00:25.324465
+2023-01-10 16:38:47.097 | INFO     | __main__:<module>:279 - execution time: 0:00:55.292320
+2023-01-10 16:41:25.711 | INFO     | __main__:<module>:229 - Analysis date: 2022-10-12
+2023-01-10 16:41:30.791 | INFO     | __main__:<module>:236 - get unique operators to loop over: 0:00:05.077763
+2023-01-10 16:42:02.094 | INFO     | __main__:<module>:276 - 110: 0:00:31.300723
+2023-01-10 16:42:33.665 | INFO     | __main__:<module>:276 - 126: 0:00:31.570437
+2023-01-10 16:43:00.439 | INFO     | __main__:<module>:276 - 127: 0:00:26.771740
+2023-01-10 16:43:20.927 | INFO     | __main__:<module>:276 - 135: 0:00:20.486447
+2023-01-10 16:43:49.134 | INFO     | __main__:<module>:276 - 148: 0:00:28.206094
+2023-01-10 16:44:09.708 | INFO     | __main__:<module>:276 - 159: 0:00:20.574076
+2023-01-10 16:44:40.772 | INFO     | __main__:<module>:276 - 167: 0:00:31.061109
+2023-01-10 16:48:53.823 | INFO     | __main__:<module>:276 - 182: 0:04:13.050599
+2023-01-10 16:49:13.052 | INFO     | __main__:<module>:276 - 188: 0:00:19.228211
+2023-01-10 16:49:55.229 | INFO     | __main__:<module>:276 - 194: 0:00:42.176460
+2023-01-10 16:50:17.054 | INFO     | __main__:<module>:276 - 218: 0:00:21.824165
+2023-01-10 16:50:36.537 | INFO     | __main__:<module>:276 - 221: 0:00:19.482938
+2023-01-10 16:51:18.017 | INFO     | __main__:<module>:276 - 226: 0:00:41.478946
+2023-01-10 16:52:41.686 | INFO     | __main__:<module>:276 - 235: 0:01:23.668528
+2023-01-10 16:53:04.147 | INFO     | __main__:<module>:276 - 243: 0:00:22.460065
+2023-01-10 16:53:26.085 | INFO     | __main__:<module>:276 - 246: 0:00:21.935803
+2023-01-10 16:53:50.366 | INFO     | __main__:<module>:276 - 247: 0:00:24.279540
+2023-01-10 16:54:15.986 | INFO     | __main__:<module>:276 - 259: 0:00:25.619674
+2023-01-10 16:54:54.567 | INFO     | __main__:<module>:276 - 269: 0:00:38.580881
+2023-01-10 16:56:15.568 | INFO     | __main__:<module>:276 - 282: 0:01:20.998230
+2023-01-10 16:56:53.530 | INFO     | __main__:<module>:276 - 284: 0:00:37.961917
+2023-01-10 16:58:00.127 | INFO     | __main__:<module>:276 - 290: 0:01:06.596406
+2023-01-10 16:58:33.224 | INFO     | __main__:<module>:276 - 293: 0:00:33.092525
+2023-01-10 17:00:01.865 | INFO     | __main__:<module>:276 - 294: 0:01:28.639597
+2023-01-10 17:00:46.232 | INFO     | __main__:<module>:276 - 295: 0:00:44.363549
+2023-01-10 17:01:17.515 | INFO     | __main__:<module>:276 - 300: 0:00:31.282114
+2023-01-10 17:01:41.962 | INFO     | __main__:<module>:276 - 301: 0:00:24.446876
+2023-01-10 17:02:02.106 | INFO     | __main__:<module>:276 - 30: 0:00:20.141867
+2023-01-10 17:02:36.297 | INFO     | __main__:<module>:276 - 310: 0:00:34.182131
+2023-01-10 17:03:13.241 | INFO     | __main__:<module>:276 - 314: 0:00:36.942048
+2023-01-10 17:03:37.295 | INFO     | __main__:<module>:276 - 315: 0:00:24.051572
+2023-01-10 17:04:19.471 | INFO     | __main__:<module>:276 - 336: 0:00:42.168281
+2023-01-10 17:04:52.190 | INFO     | __main__:<module>:276 - 349: 0:00:32.706532
+2023-01-10 17:05:23.589 | INFO     | __main__:<module>:276 - 350: 0:00:31.397957
+2023-01-10 17:05:53.810 | INFO     | __main__:<module>:276 - 361: 0:00:30.220169
+2023-01-10 17:06:16.768 | INFO     | __main__:<module>:276 - 372: 0:00:22.956382
+2023-01-10 17:06:35.742 | INFO     | __main__:<module>:276 - 45: 0:00:18.974126
+2023-01-10 17:08:52.066 | INFO     | __main__:<module>:276 - 4: 0:02:16.323434
+2023-01-10 17:09:10.348 | INFO     | __main__:<module>:276 - 75: 0:00:18.280647
+2023-01-10 17:09:10.348 | INFO     | __main__:<module>:279 - execution time: 0:27:44.635243
