--- daskify_rt/logs/A4_speeds_by_segment_trip.log
@@ -1,61 +1,0 @@
-2023-01-10 16:03:39.268 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 16:03:39.270 | INFO     | __main__:<module>:332 - start 282
-2023-01-10 16:03:39.272 | INFO     | __main__:<module>:341 - imported data: 0:00:00.000778
-2023-01-10 16:03:43.992 | INFO     | __main__:<module>:349 - keep enter/exit points by segment-trip: 0:00:04.720416
-2023-01-10 16:03:43.993 | INFO     | __main__:<module>:358 - merge in segment shapes and do linear referencing: 0:00:00.001013
-2023-01-10 16:03:43.994 | INFO     | __main__:<module>:365 - calculate speed: 0:00:00.000613
-2023-01-10 16:06:51.226 | INFO     | __main__:compute_and_export:276 - exported: 282: 0:03:07.231596
-2023-01-10 16:06:51.242 | INFO     | __main__:<module>:371 - ITP ID: 282: 0:03:11.971199
-2023-01-10 16:06:51.243 | INFO     | __main__:<module>:332 - start 4
-2023-01-10 16:06:51.244 | INFO     | __main__:<module>:341 - imported data: 0:00:00.000475
-2023-01-10 16:06:57.286 | INFO     | __main__:<module>:349 - keep enter/exit points by segment-trip: 0:00:06.042056
-2023-01-10 16:06:57.287 | INFO     | __main__:<module>:358 - merge in segment shapes and do linear referencing: 0:00:00.001389
-2023-01-10 16:06:57.288 | INFO     | __main__:<module>:365 - calculate speed: 0:00:00.000589
-2023-01-10 16:10:39.165 | INFO     | __main__:compute_and_export:276 - exported: 4: 0:03:41.876735
-2023-01-10 16:10:39.242 | INFO     | __main__:<module>:371 - ITP ID: 4: 0:03:47.998227
-2023-01-10 16:10:39.242 | INFO     | __main__:<module>:374 - execution time: 0:06:59.972535
-2023-01-10 16:03:39.268 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 16:03:39.270 | INFO     | __main__:<module>:332 - start 282
-2023-01-10 16:03:39.272 | INFO     | __main__:<module>:341 - imported data: 0:00:00.000778
-2023-01-10 16:03:43.992 | INFO     | __main__:<module>:349 - keep enter/exit points by segment-trip: 0:00:04.720416
-2023-01-10 16:03:43.993 | INFO     | __main__:<module>:358 - merge in segment shapes and do linear referencing: 0:00:00.001013
-2023-01-10 16:03:43.994 | INFO     | __main__:<module>:365 - calculate speed: 0:00:00.000613
-2023-01-10 16:06:51.226 | INFO     | __main__:compute_and_export:276 - exported: 282: 0:03:07.231596
-2023-01-10 16:06:51.242 | INFO     | __main__:<module>:371 - ITP ID: 282: 0:03:11.971199
-2023-01-10 16:06:51.243 | INFO     | __main__:<module>:332 - start 4
-2023-01-10 16:06:51.244 | INFO     | __main__:<module>:341 - imported data: 0:00:00.000475
-2023-01-10 16:06:57.286 | INFO     | __main__:<module>:349 - keep enter/exit points by segment-trip: 0:00:06.042056
-2023-01-10 16:06:57.287 | INFO     | __main__:<module>:358 - merge in segment shapes and do linear referencing: 0:00:00.001389
-2023-01-10 16:06:57.288 | INFO     | __main__:<module>:365 - calculate speed: 0:00:00.000589
-2023-01-10 16:10:39.165 | INFO     | __main__:compute_and_export:276 - exported: 4: 0:03:41.876735
-2023-01-10 16:10:39.242 | INFO     | __main__:<module>:371 - ITP ID: 4: 0:03:47.998227
-2023-01-10 16:10:39.242 | INFO     | __main__:<module>:374 - execution time: 0:06:59.972535
-2023-01-10 19:27:55.594 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 19:27:55.595 | INFO     | __main__:<module>:324 - start 182
-2023-01-10 19:27:55.595 | INFO     | __main__:<module>:333 - imported data: 0:00:00.000522
-2023-01-10 19:30:19.337 | INFO     | __main__:compute_and_export:276 - exported: 182: 0:02:23.741597
-2023-01-10 19:30:19.340 | INFO     | __main__:<module>:342 - keep enter/exit points by segment-trip: 0:02:23.744644
-2023-01-10 19:30:19.341 | INFO     | __main__:<module>:368 - ITP ID: 182: 0:02:23.746163
-2023-01-10 19:30:19.342 | INFO     | __main__:<module>:371 - execution time: 0:02:23.747236
-
-2023-01-10 16:41:34.965 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 16:41:34.966 | INFO     | __main__:<module>:324 - start 182
-2023-01-10 16:41:34.966 | INFO     | __main__:<module>:333 - imported data: 0:00:00.000541
-2023-01-10 16:43:37.032 | INFO     | __main__:<module>:341 - keep enter/exit points by segment-trip: 0:02:02.065553
-2023-01-10 16:43:37.068 | INFO     | __main__:<module>:350 - merge in segment shapes and do linear referencing: 0:00:00.036020
-2023-01-10 16:43:37.069 | INFO     | __main__:<module>:357 - calculate speed: 0:00:00.000640
-2023-01-10 19:43:27.504 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 19:43:27.506 | INFO     | __main__:<module>:324 - start 182
-2023-01-10 19:43:27.507 | INFO     | __main__:<module>:333 - imported data: 0:00:00.000600
-2023-01-10 19:43:27.508 | INFO     | __main__:<module>:341 - keep enter/exit points by segment-trip: 0:00:00.000494
-2023-01-10 19:43:28.181 | INFO     | __main__:<module>:353 - merge in segment shapes and do linear referencing: 0:00:00.673711
-2023-01-10 19:43:28.182 | INFO     | __main__:<module>:360 - calculate speed: 0:00:00.000496
-2023-01-10 19:47:05.532 | INFO     | __main__:<module>:316 - Analysis date: 2022-10-12
-2023-01-10 19:47:05.532 | INFO     | __main__:<module>:324 - start 182
-2023-01-10 19:47:05.533 | INFO     | __main__:<module>:333 - imported data: 0:00:00.000523
-2023-01-10 19:47:05.533 | INFO     | __main__:<module>:341 - keep enter/exit points by segment-trip: 0:00:00.000462
-2023-01-10 20:11:59.563 | INFO     | __main__:<module>:353 - merge in segment shapes and do linear referencing: 0:24:54.029834
-2023-01-10 20:11:59.564 | INFO     | __main__:<module>:360 - calculate speed: 0:00:00.001013
-2023-01-10 22:37:28.002 | INFO     | __main__:compute_and_export:276 - exported: 182: 2:25:28.437116
-2023-01-10 22:37:28.062 | INFO     | __main__:<module>:366 - ITP ID: 182: 2:50:22.529838
-2023-01-10 22:37:28.063 | INFO     | __main__:<module>:369 - execution time: 2:50:22.530679
